Moving everything to proper files
