+++ resources/assets/scss/app.scss
@@ -20,6 +20,7 @@
 @import "_components/_facebook-login";
 @import "_components/_cover-image";
 @import "_components/_divider";
+@import "_components/_password-visibility";
 
 // Regions - complete sections of an interface
 @import "_regions/_chrome";
